B[MAINTENANCE] update api docs

+++ models/category.go
@@ -7,7 +7,7 @@ type Category struct {
 	// 分类类型，0代表菜品，1代表套餐;默认是菜品
 	Type uint8 `gorm:"default:0;not null" form:"type" validate:"gte=0,lte=1" json:"type"`
 	// 排序值，用于决定顺序；
-	Sort uint `gorm:"default:0;not null" form:"sort" validate:"gte=0" json:"sort"`
+	Sort uint16 `gorm:"default:0;not null" form:"sort" validate:"gte=0" json:"sort"`
 	// 0代表禁用，1代表启用，默认禁用
 	Status uint8 `gorm:"default:0;not null" form:"status" validate:"gte=0,lte=1" json:"status"`
 	// 考虑到每一家店铺大概率有一个独立的分类，因此每一个category对应一家店铺
